[Improvement] Improved inline conditions
